Add the VPA for OpenSRP server web chart
Add the Vertical Pod Autoscaler to the OpenSRP Server Web Helm chart to
help with figuring out what the resource requests and limits should be
for the service.

By default, the VPA will be enabled in suggestion mode to allow people
view the suggestions (which won't be auto-applied).

Effort: 1:00 hours

Signed-off-by: Jason Rogena <jason@rogena.me>