Add the VPA to all other charts
Add the Vertical Pod Autoscaler to all the charts except the OpenSRP
Server Web Helm chart (which already has it) to help with figuring out
what the resource requests and limits should for the services.

By default, the VPA will be enabled in suggestion mode to allow people
view the suggestions (which won't be auto-applied).

Effort: 0:30 hours

Signed-off-by: Jason Rogena <jason@rogena.me>